requirements.txt
# Speech-to-Text (Whisper)
openai-whisper==20231117
whisper-cpp==1.2.0

# Translation
transformers==4.36.0
torch==2.1.0
sacrebleu==2.3.1

# Text-to-Speech (Coqui TTS)
TTS==0.22.0
coqui-tts==0.22.0

# Audio processing
pyaudio==0.2.11
soundfile==0.12.1
numpy==1.24.3

# Additional utilities
scipy==1.11.4
librosa==0.10.1
sentencepiece==0.1.99
protobuf==4.25.1

# Web API
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# TTS runtime dependencies
gTTS==2.5.1
pygame==2.6.1
